+++ Sources/Sentry/Public/SentryOptions.h
@@ -17,9 +17,10 @@ NS_SWIFT_NAME(Options)
                       didFailWithError:(NSError *_Nullable *_Nullable)error;
 
 /**
- * Optional, custom cache directory. Use when default one can not be accessed, e.g. in security environment.
+ * Optional, custom cache directory. Use when default one can not be accessed, e.g. in security
+ * environment.
  */
-@property (nonatomic, strong) NSString *_Nullable customCacheDirectory;
+@property (nonatomic, strong) NSString *_Nullable cacheDirectory;
 
 /**
  * The DSN tells the SDK where to send the events to. If this value is not provided, the SDK will