+++ Sources/Sentry/SentryCrashIntegration.m
@@ -81,8 +81,7 @@ SentryCrashIntegration ()
     self.scopeObserver =
         [[SentryCrashScopeObserver alloc] initWithMaxBreadcrumbs:options.maxBreadcrumbs];
 
-    NSString *customDirectory = options.customCacheDirectory;
-    [self startCrashHandler:customDirectory];
+    [self startCrashHandler:options.cacheDirectory];
 
     if (options.stitchAsyncCode) {
         [self.crashAdapter installAsyncHooks];
@@ -98,7 +97,7 @@ SentryCrashIntegration ()
     return kIntegrationOptionEnableCrashHandler;
 }
 
-- (void)startCrashHandler:(NSString *)customCacheDirectory
+- (void)startCrashHandler:(NSString *)cacheDirectory
 {
     void (^block)(void) = ^{
         BOOL canSendReports = NO;
@@ -115,7 +114,7 @@ SentryCrashIntegration ()
             canSendReports = YES;
         }
 
-        [installation install:customCacheDirectory];
+        [installation install:cacheDirectory];
 
         // We need to send the crashed event together with the crashed session in the same envelope
         // to have proper statistics in release health. To achieve this we need both synchronously